+++ weatherDB/utils/dwd.py
@@ -343,7 +343,7 @@ def get_dwd_meta(ftp_folder):
 
     # trim whitespace in string columns
     for dtype, col in zip(meta.dtypes, meta.columns):
-        if pd.api.types.is_string_dtype(dtype):
+        if pd.api.types.is_string_dtype(dtype) and col != "geometry":
             meta[col] = meta[col].str.strip()
 
     # return